Fix relative path issue
When you use sys.path += ["../modules"] to add a path to the sys.path list, the path is added relative to the current working directory of the script at the time the code is executed.

For example, if your script is located in the directory /home/user/project and you run the script from that directory, then sys.path += ["../modules"] will add /home/user/modules to the sys.path list. However, if you run the script from a different directory, the relative path will be resolved relative to that directory, and the path added to sys.path will not be correct.

A better way to add a relative path is to use `os.path' module as commited.

A more convenient way is to allow your code to be able to see a user-defined package: for example, from cs285 hw1:

- define setup.py which contains

	# setup.py
	from setuptools import setup

	setup(
	    name='cs285',
	    version='0.1.0',
	    packages=['cs285'],
	)

- then install

	```
	cd <path_to_hw1>
	$ pip install -e .
	```

then, one can simply import cs285 and its relative paths as

	```
	import cs285.infrastructure
	```

This might be the best practice, and will be implemented later.